refactor: consolidate villager commands into single command
Consolidate scattered villager debug commands by removing obsolete ones and merging functionality into the main villager command.

Changes:
- Remove /vw assign command (will be replaced with zone-based assignment)
- Remove /vw exhaust command (functionality merged into hunger command)
- Remove /vw hunger command (functionality moved to villager command)
- Add /vw villager <targets> hunger subcommand to display hunger status
- Add /vw villager <targets> hunger <exhaustion> to add exhaustion for testing

Resolves #41
